Implement nested resource data structures
After a lot of consideration, I have decided to use a REST-like
resource structure for requests and responses. To minimize the
number of classes, the same resource class is used both for request
and response data. I hope that this structure will be more useful
when developing the frontend. The tests are also more elegant.

I consider 'flat' data superior in some ways because it is
stricter and much more consistent. However since this API is
targeted at JavaScript, an easy-to-use structure seems appropriate.

Note that the resource data is generated from the SessionData, which
contains all retrieved database rows. This means it is still very
easy to generate flat data instead.