Remove categories completely
This is quite a big change. The idea of categories has been part
of the database for quite a while. However I realized it does not
actually add much value. Meanwhile it does complicate the design.
The categories where primarily invented to be able to separate
definitions. However, this is not actually necessary since the
function IDs itself provide enough room for separation.