Rename locale to language
I decided the usage of 'locale' is wrong. A locale is much more than
just a language. The upside is that locale is shorter and less
commonly used than 'language' (causing less conflicts, and in this
case making refactoring super easy). But I find correctness and
clarity more important in this case.