Refactor and fix bug with eqlib
EqLib computed wrong rearrangement positions.
Also, eqlib can produce rearrangements that do not change the
expression (when arguments with a different rearrangable function
are different but not the expression itself). This is a bit tedious
to fix. But we can simply ignore steps that do not change the
expression. Also this would ignore substitutions that do not change
anything which is probably a good thing.